added wave efect on water, some refactoring

+++ src/ship.hpp
@@ -8,12 +8,12 @@ private:
 	int size;
 	int num_of_hit_fields;
 	bool alive;
-	int pos_x, pos_y;
+	float pos_x, pos_y;
 	bool orientation;
 
 	void update_status();
 public:
-	ship(int size, int pos_x, int pos_y, bool orientation);
+	ship(int size, float pos_x, float pos_y, bool orientation);
 	int get_size(void);
 	bool get_status(void);
 	void draw();
